docs: use react template for table
