+++ components/table/web-component/story-template.jsx
@@ -1,0 +1,32 @@
+/**
+ * @license EUPL-1.2
+ * Copyright (c) 2022 Gemeente Utrecht
+ * Copyright (c) 2022 Frameless B.V.
+ */
+import React from 'react';
+
+export const Table = () => (
+  <utrecht-table aria-labelledby="53818ecc-f9a6-4a9d-9e63-12a5ebb89e31">
+    <utrecht-table-caption id="53818ecc-f9a6-4a9d-9e63-12a5ebb89e31">Table</utrecht-table-caption>
+    <utrecht-table-header>
+      <utrecht-table-row>
+        <utrecht-table-header-cell scope="col">Column header cell 1</utrecht-table-header-cell>
+        <utrecht-table-header-cell scope="col">Column header cell 2</utrecht-table-header-cell>
+      </utrecht-table-row>
+    </utrecht-table-header>
+    <utrecht-table-footer>
+      <utrecht-table-row>
+        <utrecht-table-cell>Footer cell 1</utrecht-table-cell>
+        <utrecht-table-cell>Footer cell 2</utrecht-table-cell>
+      </utrecht-table-row>
+    </utrecht-table-footer>
+    <utrecht-table-body>
+      <utrecht-table-row>
+        <utrecht-table-header-cell scope="row">Row header cell</utrecht-table-header-cell>
+        <utrecht-table-cell>Cell</utrecht-table-cell>
+      </utrecht-table-row>
+    </utrecht-table-body>
+  </utrecht-table>
+);
+
+export default Table;
