--- components/table/web-component/template.js
@@ -1,23 +1,0 @@
-export const TableExample = () => `
-<utrecht-table aria-labelledby="53818ecc-f9a6-4a9d-9e63-12a5ebb89e31">
-  <utrecht-table-caption id="53818ecc-f9a6-4a9d-9e63-12a5ebb89e31">Table</utrecht-table-caption>
-  <utrecht-table-header>
-    <utrecht-table-row>
-      <utrecht-table-header-cell scope="col">Column header cell 1</utrecht-table-header-cell>
-      <utrecht-table-header-cell scope="col">Column header cell 2</utrecht-table-header-cell>
-    </utrecht-table-row>
-  </utrecht-table-header>
-  <utrecht-table-footer>
-    <utrecht-table-row>
-      <utrecht-table-cell>Footer cell 1</utrecht-table-cell>
-      <utrecht-table-cell>Footer cell 2</utrecht-table-cell>
-    </utrecht-table-row>
-  </utrecht-table-footer>
-  <utrecht-table-body>
-    <utrecht-table-row>
-      <utrecht-table-header-cell scope="row">Row header cell</utrecht-table-header-cell>
-      <utrecht-table-cell>Cell</utrecht-table-cell>
-    </utrecht-table-row>
-  </utrecht-table-body>
-</utrecht-table>
-`;
